+++ CHANGELOG.md
@@ -59,6 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spelling of `ArticulatedPart::Get/SetAttachmentID` methods via [#71](https://github.com/karljj1/kdis/pull/71)
 - Type of `Entity_State_PDU::ApplyDeadReckoning`'s parameter to KFLOAT32 via [#71](https://github.com/karljj1/kdis/pull/71)
 - Spelling of `KDIS::UTILS::Matrix::inPlaceTranspose()` method via [#71](https://github.com/karljj1/kdis/pull/71)
+- Spelling of `KDIS::DATA_TYPE::EntityAppearance` "GuidedMunitions" methods via [#71](https://github.com/karljj1/kdis/pull/71)
 
 ### Security